move desugaring to item bounds
